# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Muse API",
    "description": "Muse Music Streaming Service API Documentation",
    "contact": {
      "email": "mubareksd@gmail.com"
    },
    "license": {
      "name": "GPL 3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://api.muse.com/v1"
    }
  ],
...
  "paths": {
...
    "/albums": {
      "get": {
        "tags": ["Albums"],
        "operationId": "getseveralalbums",
        "summary": "Albums",
        "description": "Get Several Albums",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "example": 10,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "offset",
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeveralAlbumsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/albums/{id}": {
      "get": {
        "tags": ["Albums"],
        "operationId": "getalbumbyid",
        "summary": "Albums",
        "description": "Get Album with id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "8fa3f58b-6e88-4d7c-9b7e-aacbdd9a77fe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Albums"],
        "operationId": "updatealbum",
        "summary": "Albums",
        "description": "Update Album",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "8fa3f58b-6e88-4d7c-9b7e-aacbdd9a77fe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbum"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Successfully"
          }
        }
      }
    },
...
    "/artists/{id}/albums": {
      "get": {
        "tags": ["Artists"],
        "summary": "Artists",
        "description": "Get Artists Albums",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "e242bc9a0ceea852400d3f927f07557e4a36b9f34e7f",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "example": 10,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "offset",
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      },
      "post": {
        "tags": ["Artists"],
        "summary": "Artists",
        "description": "Create Artist Album",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "e242bc9a0ceea852400d3f927f07557e4a36b9f34e7f",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistAlbums"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created Successfully"
          }
        }
      },
      "delete": {
        "tags": ["Artists"],
        "summary": "Artists",
        "description": "Delete Artist Album",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "e242bc9a0ceea852400d3f927f07557e4a36b9f34e7f",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistAlbums"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deleted Successfully"
          }
        }
      }
    },
...
    "/me/albums": {
      "get": {
        "tags": ["Me"],
        "summary": "Me",
        "description": "Get Current users Albums",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SeveralAlbumsResponse": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "example": 10,
            "description": "Number of Albums"
          },
          "offset": {
            "type": "number",
            "example": 0,
            "description": "Number of Albums"
          },
          "total": {
            "type": "number",
            "example": 1,
            "description": "Number of Albums"
          },
          "albums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Album"
            }
          }
        }
      },
      "Album": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "41d0407b-c0cb-483e-a047-b6a6f4cb1620",
            "description": "Album ID"
          },
          "name": {
            "type": "string",
            "example": "The Dark Side of the Moon",
            "description": "Album name"
          },
          "type": {
            "type": "string",
            "example": "ALBUM",
            "description": "Album type"
          },
          "release_date": {
            "type": "string",
            "example": "1973-03-01",
            "description": "Album release date"
          }
        }
      },
      "UpdateAlbum": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "The Dark Side of the Moon",
            "description": "Album name"
          },
          "type": {
            "type": "string",
            "example": "ALBUM",
            "description": "Album type"
          },
          "release_date": {
            "type": "string",
            "example": "1973-03-01",
            "description": "Album release date"
          }
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}